Final Version Of The Game
The game is fully completed and the animations for the game battle are all in place. Possible improvements is a hunt feature, which will activate when the enemy hits a ship and then it will try to hit tiles that are close to that area in a systematic way to fully sink the ship. Another possible future is the implementation of score and the addition of a database, which will save those scores for the user. Some animations could also improve. That being said I will leave it as it is, because I am out of time.